samples/12-sharded-catalog/requirements.txt
# Sample requirements - using ToolWeaver from PyPI
toolweaver[vector]==0.1.3

# Environment variables
python-dotenv>=1.0.0
